C4-71 complete initial refactor for filter sets
