+++ src/encoded/__init__.py
@@ -244,7 +244,7 @@ def main(global_config, **local_config):
 
     if 'elasticsearch.server' in config.registry.settings:
         config.include('snovault.elasticsearch')
-        config.include('.search')
+        config.include('.search.search')
 
     # this contains fall back url, so make sure it comes just before static_resoruces
     config.include('.types.page')
